Support CREATE OR REPLACE in table materialization
