+++ .changes/unreleased/Features-20231129-195500.yaml
@@ -1,0 +1,7 @@
+kind: Features
+body: Support CREATE OR REPLACE in table materialization
+time: 2023-11-29T19:55:00.460963+01:00
+custom:
+  Author: damian3031
+  Issue: "371"
+  PR: "377"
